Separate decorators from denguebot_fsm
